requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.98.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx>=0.25.0
websockets>=12.0 